requirements.txt
fastapi
uvicorn
peewee
mysqlclient
python-dotenv